+++ platform/velocity/src/main/java/net/silthus/schat/velocity/adapter/VelocityMessagingService.java
@@ -22,25 +22,14 @@
  *  SOFTWARE.
  */
 
-package net.silthus.schat.bukkit.adapter;
+package net.silthus.schat.velocity.adapter;
 
-import net.silthus.schat.eventbus.AbstractEventBus;
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.Listener;
-import org.bukkit.event.server.PluginDisableEvent;
-import org.bukkit.plugin.Plugin;
+import net.silthus.schat.messaging.Messenger;
 
-public final class BukkitEventBus extends AbstractEventBus<Plugin> implements Listener {
+public class VelocityMessagingService implements Messenger.Gateway {
+
     @Override
-    protected Plugin checkPlugin(Object plugin) throws IllegalArgumentException {
-        if (plugin instanceof Plugin p)
-            return p;
-        else
-            throw new IllegalArgumentException();
-    }
+    public void sendOutgoingMessage(String encodedMessage) {
 
-    @EventHandler
-    public void onPluginDisable(PluginDisableEvent event) {
-        unregisterHandlers(event.getPlugin());
     }
 }